requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==13.1
redis==5.2.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
numpy==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
deepgram-sdk==3.8.0
langdetect==1.0.9
orjson==3.10.13
soundfile==0.12.1
aiofiles==24.1.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.25.0